New endpoints, see GraphQlSchemaDefinition or query the schema

+++ gui/app/models/cskg/Edge.scala
@@ -6,5 +6,5 @@ final case class Edge(
                       other: Option[String], // No default values so the compiler can check missing fields on construction
                       predicate: String,
                       subject: String,
-                      weight: Option[Float]
+                      weight: Option[Double]
                      )
